# spikes/service-generation-spike/openapi/gen/go-gin-server/go/model_error.go
/*
 * Swagger Petstore
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Error struct {
	Code int32 `json:"code"`

	Message string `json:"message"`
}